political_leaning_bot/train_model.py
import pandas as pd
from sklearn.svm import LinearSVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
import joblib

# Load dataset
data = pd.read_csv('dataset.csv')
X = data[[f'q{i+1}' for i in range(20)]]
y = data['label']

# Split data
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Train model
model = LinearSVC(max_iter=2000)
model.fit(X_train, y_train)

# Evaluate
predictions = model.predict(X_test)
print(f"Accuracy: {accuracy_score(y_test, predictions):.2f}")

# Save model
joblib.dump(model, 'political_leaning/model.sav')
print("Model saved to political_leaning/model.sav")