political_leaning_bot/requirements.txt
django==4.2 
djangorestframework==3.15 
scikit-learn==1.3 
gunicorn==23.0.0 
pandas==2.2 
numpy==1.26.4 
joblib==1.4 
psycopg==3.2 
dj-database-url==2.1.0